Added Weapon for Demo + shooting
Added a small weapons art pack (temporary) to show off for demo. Added to my prefabs, art, and scripts folder to create a gun that rotates towards the mouse and shoots bullets.  On impact, particle system is created and played.

Also fixed the issue with the main player not being able to be seen. And added my gun prefab, as a child of the player in the scene, so that we can show it off.