Added Damage function to BulletScript
Added a function to the script which is placed on bullets, which (hopefully) damages an enemy, if it collides with an enemy.

The health of an enemy, and functions to damage it/"kill" it are assumed to be on the enemy. This script only calls a "damage" function of the component.

Examples of what I'm talking about are posted in the discord