Added generic classes for bullet,impact,gun
Changed the format of my scripts to be the baseline, generic class scripts for guns, bullets, and bullet impacts. Then created scripts for the specific non-default weapon, bullet, and impact. However, these are left blank for now. But any further changes can be made to any script which has a class that inherits from the main.